requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-cognitiveservices-speech==1.23.0
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==4.4.2
idna==3.4
imageio==2.22.0
imageio-ffmpeg==0.4.7
moviepy==1.0.0
numpy==1.23.3
Pillow==9.2.0
proglog==0.1.10
requests==2.28.1
tqdm==4.64.1
urllib3==1.26.12
azure-storage-blob
